+++ src/dao/GlassDAO.js
@@ -3,7 +3,7 @@ import type { Glass } from '../types';
 
 import DAO from './DAO';
 import { DAO_ENTITIES } from '../constants';
-import DefaultTranslator from '../translators/Default';
+import DefaultTranslator from '../translators/DefaultTranslator';
 
 class GlassDAO extends DAO<Glass, Glass> {
   constructor() {
