+++ src/dao/LocationDAO.js
@@ -3,7 +3,7 @@ import type { Location } from '../types';
 
 import DAO from './DAO';
 import { DAO_ENTITIES } from '../constants';
-import DefaultTranslator from '../translators/Default';
+import DefaultTranslator from '../translators/DefaultTranslator';
 
 class LocationDAO extends DAO<Location, Location> {
   constructor() {
