+++ src/dao/SrmDAO.js
@@ -3,7 +3,7 @@ import type { Srm } from '../types';
 
 import DAO from './DAO';
 import { DAO_ENTITIES } from '../constants';
-import DefaultTranslator from '../translators/Default';
+import DefaultTranslator from '../translators/DefaultTranslator';
 
 class SrmDAO extends DAO<Srm, Srm> {
   constructor() {
