+++ src/dao/StyleDAO.js
@@ -3,7 +3,7 @@ import type { Style } from '../types';
 
 import DAO from './DAO';
 import { DAO_ENTITIES } from '../constants';
-import DefaultTranslator from '../translators/Default';
+import DefaultTranslator from '../translators/DefaultTranslator';
 
 class StyleDAO extends DAO<Style, Style> {
   constructor() {
