+++ src/dao/TapDAO.js
@@ -3,7 +3,7 @@ import type { Tap, TapMutator } from '../types';
 
 import DAO from './DAO';
 import { DAO_ENTITIES } from '../constants';
-import TapTranslator from '../translators/Tap';
+import TapTranslator from '../translators/TapTranslator';
 
 class TapDAO extends DAO<Tap, TapMutator> {
   constructor() {
