+++ src/translators/DefaultTranslator.js
